+++ CHANGELOG.md
@@ -1,5 +1,10 @@
 ## Changelog
 
+### 12.64.0
+
+* `[timeutil]` Added method `MiniDuration` which returns formatted value for short durations (_e.g. s/ms/us/ns_)
+* `[terminal]` Method `ReadUI` marked as deprecated (_use method `Read` instead_)
+
 ### 12.63.0
 
 * `[knf]` Fixed bug with using method `Is` for checking for empty values